Diversas mudanças e refatorações
Removidas classes módulos supérfluos:
    - Classe Collider
    - Classe StageState
    - Módulo Collision.cpp

Implementada cena inicial (somente texto) na classe IntroState;
    - Ainda é necessário implementar carregamento de HUD.

Criada a classe ScrollerText para facilitar lógica de scrolling e identificação de texto;
    - Scrolling ainda não funciona com roda do mouse - implementado por setas.

Adicionado membro "id" na classe Button;
    - Forma melhor de implementar diferenciação para requisição de botões;
    - Novo construtor adicionado para permitir instanciação sem identificação.

Classe ExploreState modificada para mostrar cena inicial no começo da execução;

Novos arquivos de texto incluídos;

Mudanças de refatoração em IntroState e ExploreState:
    - Trechos de criação de Sprites, Texts e Buttons refatorados em funções novas, agora presentes no módulo "GeneralFunctions".
    - Refatoração ainda precisa ser aplicada em:
        - CombatState;
        - TitleState.

+++ Rest in Peace/Rect.cpp
@@ -96,7 +96,7 @@ void Rect::ResizeThis(float ratio) {
  * Seta posição do retângulo para que centro esteja no ponto fornecido.
  */
 void Rect::SetCenterPosition(Vec2 vector) {
-	Vec2* offset = new Vec2(w/2, h/2);
+	Vec2* offset = new Vec2((int)(w/2), (int)(h/2));
 	offset->RotateThis(angle);
 	SetPosition(vector - *offset);
 }
